% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_saving.R
\name{copy_images_to_class_folders}
\alias{copy_images_to_class_folders}
\title{Copy images to class-organized folders}
\usage{
copy_images_to_class_folders(
  classifications,
  src_folder,
  temp_folder,
  output_folder
)
}
\arguments{
\item{classifications}{Classifications data frame with file_name and class_name columns}

\item{src_folder}{Source folder containing PNG images}

\item{temp_folder}{Temporary folder for ifcb_annotate_samples}

\item{output_folder}{Permanent output folder for class-organized images}
}
\value{
NULL (called for side effects)
}
\description{
Copies PNG images from a flat source folder into class-organized subfolders,
both for temporary use by ifcb_annotate_samples and for permanent storage.
}
\examples{
\dontrun{
# Copy images to class folders
copy_images_to_class_folders(
  classifications = data.frame(
    file_name = c("sample_00001.png", "sample_00002.png"),
    class_name = c("Diatom", "Ciliate")
  ),
  src_folder = "/tmp/png/sample",
  temp_folder = "/tmp/annotate",
  output_folder = "/data/png_output"
)
}
}
