% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instrument_profile.R
\name{DEFAULT_IMAGE_EXTENSIONS}
\alias{DEFAULT_IMAGE_EXTENSIONS}
\title{Default image file extensions for generic mode}
\usage{
DEFAULT_IMAGE_EXTENSIONS
}
\description{
Default image file extensions for generic mode
}
\keyword{internal}
