% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_loading.R
\name{filter_to_extracted}
\alias{filter_to_extracted}
\title{Filter classifications to only include extracted images}
\usage{
filter_to_extracted(classifications, extracted_folder)
}
\arguments{
\item{classifications}{Data frame of classifications (must have file_name column)}

\item{extracted_folder}{Path to folder with extracted image files}
}
\value{
Filtered classifications data frame
}
\description{
Filters a classifications data frame to only include ROIs that have
corresponding image files in the extracted folder. Matching is by file name,
so it works for IFCB PNG exports as well as generic image files (jpg, etc.).
}
\examples{
\dontrun{
# Filter to only images that were successfully extracted
classifications <- filter_to_extracted(
  classifications = classifications,
  extracted_folder = "/tmp/png/D20230101T120000_IFCB134"
)
}
}
