% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{get_db_path}
\alias{get_db_path}
\title{Get path to the annotations SQLite database}
\usage{
get_db_path(db_folder)
}
\arguments{
\item{db_folder}{Path to the database directory. Defaults to
\code{\link{get_default_db_dir}()}, a persistent local directory.}
}
\value{
Path to the SQLite database file
}
\description{
Returns the path to \code{annotations.sqlite} in the given database
directory. The database directory should be on a local filesystem, not a
network drive, because
\href{https://www.sqlite.org/useovernet.html}{SQLite file locking is
unreliable over network filesystems}.
}
\examples{
# Use the default local database directory
get_db_path(get_default_db_dir())

# Or specify a custom directory
get_db_path("/data/local_db")
}
\seealso{
\code{\link{get_default_db_dir}} for the default database directory
}
