% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_file_index_path}
\alias{get_file_index_path}
\title{Get path to file index cache}
\usage{
get_file_index_path()
}
\value{
Path to the file index JSON file
}
\description{
Returns the path to the file index JSON cache file. The file index
stores scanned folder results to avoid expensive recursive directory
scans on startup.
}
