% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instrument_profile.R
\name{INSTRUMENT_TYPES}
\alias{INSTRUMENT_TYPES}
\title{Supported instrument profile types}
\usage{
INSTRUMENT_TYPES
}
\value{
Character vector of valid \code{instrument_type} values.
}
\description{
Supported instrument profile types
}
\keyword{internal}
