% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{list_annotated_samples_db}
\alias{list_annotated_samples_db}
\title{List samples with annotations in the database}
\usage{
list_annotated_samples_db(db_path)
}
\arguments{
\item{db_path}{Path to the SQLite database file}
}
\value{
Character vector of sample names that have annotations
}
\description{
List samples with annotations in the database
}
\examples{
\dontrun{
db_path <- get_db_path("/data/manual")
samples <- list_annotated_samples_db(db_path)
}
}
