% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{save_file_index}
\alias{save_file_index}
\title{Save file index to disk cache}
\usage{
save_file_index(data)
}
\arguments{
\item{data}{List containing scan results (sample names, path maps, etc.)}
}
\value{
NULL (called for side effects)
}
\description{
Writes the file index data to a JSON cache file for fast startup.
}
